requirements.txt
django == 2.2.3
django-material == 1.4.3
Pillow == 5.00
django-imagekit == 4.0.2
django-leaflet == 0.25.0
django-geojson == 2.12.0
jsonfield == 2.0.2
django-import-export == 1.2.0
psycopg2 == 2.8.4